Fix warning: format ‘%i’ expects argument of type ‘int’, but argument 3 has type ‘integer’

+++ 3rdparty/clapack/src/xerbla.c
@@ -56,7 +56,7 @@ static integer c__1 = 1;
 /*     .. Executable Statements .. */
 
     printf("** On entry to %6s, parameter number %2i had an illegal value\n",
-		srname, *info);
+		srname, (int)(*info));
 
 
 /*     End of XERBLA */
